HYD-4648 Create ExtractAPI module
A common operation in the gui is to extract the id from an API call.
Such an example might be a call to "/api/hosts/10". In this case, the id is 10.

Since this is a common operation we would like to encapsulate the logic to
extract the id from the path.
In this patch we will add a new module called "extract-api" under the
ui-modules/isomorphic folder.

  - Added index.js as main module file
  - Added test files
  - Updated cluster_setup to include extract-api dependencies
  - Update run_tests to run the extract-api tests
  - Updated test_dependencies to include extract-api
  - Added package.json to define module and dependencies

Signed-off-by: Will Johnson <william.c.johnson@intel.com>